jooq config
