--- src/main/java/eu/dissco/backend/database/jooq/DefaultCatalog.java
@@ -1,44 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package eu.dissco.backend.database.jooq;
-
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.jooq.Schema;
-import org.jooq.impl.CatalogImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class DefaultCatalog extends CatalogImpl {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * The reference instance of <code>DEFAULT_CATALOG</code>
-     */
-    public static final DefaultCatalog DEFAULT_CATALOG = new DefaultCatalog();
-
-    /**
-     * The schema <code>public</code>.
-     */
-    public final Public PUBLIC = Public.PUBLIC;
-
-    /**
-     * No further instances allowed
-     */
-    private DefaultCatalog() {
-        super("");
-    }
-
-    @Override
-    public final List<Schema> getSchemas() {
-        return Arrays.<Schema>asList(
-            Public.PUBLIC);
-    }
-}
