--- src/main/java/eu/dissco/backend/database/jooq/Keys.java
@@ -1,32 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package eu.dissco.backend.database.jooq;
-
-
-import eu.dissco.backend.database.jooq.tables.NewMapping;
-import eu.dissco.backend.database.jooq.tables.NewSourceSystem;
-import eu.dissco.backend.database.jooq.tables.records.NewMappingRecord;
-import eu.dissco.backend.database.jooq.tables.records.NewSourceSystemRecord;
-
-import org.jooq.TableField;
-import org.jooq.UniqueKey;
-import org.jooq.impl.DSL;
-import org.jooq.impl.Internal;
-
-
-/**
- * A class modelling foreign key relationships and constraints of tables in 
- * public.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Keys {
-
-    // -------------------------------------------------------------------------
-    // UNIQUE and PRIMARY KEY definitions
-    // -------------------------------------------------------------------------
-
-    public static final UniqueKey<NewMappingRecord> NEW_MAPPING_PK = Internal.createUniqueKey(NewMapping.NEW_MAPPING, DSL.name("new_mapping_pk"), new TableField[] { NewMapping.NEW_MAPPING.ID, NewMapping.NEW_MAPPING.VERSION }, true);
-    public static final UniqueKey<NewSourceSystemRecord> NEW_SOURCE_SYSTEM_ENDPOINT_KEY = Internal.createUniqueKey(NewSourceSystem.NEW_SOURCE_SYSTEM, DSL.name("new_source_system_endpoint_key"), new TableField[] { NewSourceSystem.NEW_SOURCE_SYSTEM.ENDPOINT }, true);
-    public static final UniqueKey<NewSourceSystemRecord> NEW_SOURCE_SYSTEM_PKEY = Internal.createUniqueKey(NewSourceSystem.NEW_SOURCE_SYSTEM, DSL.name("new_source_system_pkey"), new TableField[] { NewSourceSystem.NEW_SOURCE_SYSTEM.ID }, true);
-}
