--- src/main/java/eu/dissco/backend/database/jooq/Tables.java
@@ -1,26 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package eu.dissco.backend.database.jooq;
-
-
-import eu.dissco.backend.database.jooq.tables.NewMapping;
-import eu.dissco.backend.database.jooq.tables.NewSourceSystem;
-
-
-/**
- * Convenience access to all tables in public.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Tables {
-
-    /**
-     * The table <code>public.new_mapping</code>.
-     */
-    public static final NewMapping NEW_MAPPING = NewMapping.NEW_MAPPING;
-
-    /**
-     * The table <code>public.new_source_system</code>.
-     */
-    public static final NewSourceSystem NEW_SOURCE_SYSTEM = NewSourceSystem.NEW_SOURCE_SYSTEM;
-}
