--- src/main/java/eu/dissco/backend/database/jooq/tables/records/NewSourceSystemRecord.java
@@ -1,331 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package eu.dissco.backend.database.jooq.tables.records;
-
-
-import eu.dissco.backend.database.jooq.tables.NewSourceSystem;
-
-import java.time.Instant;
-
-import org.jooq.Field;
-import org.jooq.Record1;
-import org.jooq.Record7;
-import org.jooq.Row7;
-import org.jooq.impl.UpdatableRecordImpl;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class NewSourceSystemRecord extends UpdatableRecordImpl<NewSourceSystemRecord> implements Record7<String, String, String, String, Instant, Instant, String> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * Setter for <code>public.new_source_system.id</code>.
-     */
-    public void setId(String value) {
-        set(0, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.id</code>.
-     */
-    public String getId() {
-        return (String) get(0);
-    }
-
-    /**
-     * Setter for <code>public.new_source_system.name</code>.
-     */
-    public void setName(String value) {
-        set(1, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.name</code>.
-     */
-    public String getName() {
-        return (String) get(1);
-    }
-
-    /**
-     * Setter for <code>public.new_source_system.endpoint</code>.
-     */
-    public void setEndpoint(String value) {
-        set(2, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.endpoint</code>.
-     */
-    public String getEndpoint() {
-        return (String) get(2);
-    }
-
-    /**
-     * Setter for <code>public.new_source_system.description</code>.
-     */
-    public void setDescription(String value) {
-        set(3, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.description</code>.
-     */
-    public String getDescription() {
-        return (String) get(3);
-    }
-
-    /**
-     * Setter for <code>public.new_source_system.created</code>.
-     */
-    public void setCreated(Instant value) {
-        set(4, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.created</code>.
-     */
-    public Instant getCreated() {
-        return (Instant) get(4);
-    }
-
-    /**
-     * Setter for <code>public.new_source_system.deleted</code>.
-     */
-    public void setDeleted(Instant value) {
-        set(5, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.deleted</code>.
-     */
-    public Instant getDeleted() {
-        return (Instant) get(5);
-    }
-
-    /**
-     * Setter for <code>public.new_source_system.mapping_id</code>.
-     */
-    public void setMappingId(String value) {
-        set(6, value);
-    }
-
-    /**
-     * Getter for <code>public.new_source_system.mapping_id</code>.
-     */
-    public String getMappingId() {
-        return (String) get(6);
-    }
-
-    // -------------------------------------------------------------------------
-    // Primary key information
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Record1<String> key() {
-        return (Record1) super.key();
-    }
-
-    // -------------------------------------------------------------------------
-    // Record7 type implementation
-    // -------------------------------------------------------------------------
-
-    @Override
-    public Row7<String, String, String, String, Instant, Instant, String> fieldsRow() {
-        return (Row7) super.fieldsRow();
-    }
-
-    @Override
-    public Row7<String, String, String, String, Instant, Instant, String> valuesRow() {
-        return (Row7) super.valuesRow();
-    }
-
-    @Override
-    public Field<String> field1() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.ID;
-    }
-
-    @Override
-    public Field<String> field2() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.NAME;
-    }
-
-    @Override
-    public Field<String> field3() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.ENDPOINT;
-    }
-
-    @Override
-    public Field<String> field4() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.DESCRIPTION;
-    }
-
-    @Override
-    public Field<Instant> field5() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.CREATED;
-    }
-
-    @Override
-    public Field<Instant> field6() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.DELETED;
-    }
-
-    @Override
-    public Field<String> field7() {
-        return NewSourceSystem.NEW_SOURCE_SYSTEM.MAPPING_ID;
-    }
-
-    @Override
-    public String component1() {
-        return getId();
-    }
-
-    @Override
-    public String component2() {
-        return getName();
-    }
-
-    @Override
-    public String component3() {
-        return getEndpoint();
-    }
-
-    @Override
-    public String component4() {
-        return getDescription();
-    }
-
-    @Override
-    public Instant component5() {
-        return getCreated();
-    }
-
-    @Override
-    public Instant component6() {
-        return getDeleted();
-    }
-
-    @Override
-    public String component7() {
-        return getMappingId();
-    }
-
-    @Override
-    public String value1() {
-        return getId();
-    }
-
-    @Override
-    public String value2() {
-        return getName();
-    }
-
-    @Override
-    public String value3() {
-        return getEndpoint();
-    }
-
-    @Override
-    public String value4() {
-        return getDescription();
-    }
-
-    @Override
-    public Instant value5() {
-        return getCreated();
-    }
-
-    @Override
-    public Instant value6() {
-        return getDeleted();
-    }
-
-    @Override
-    public String value7() {
-        return getMappingId();
-    }
-
-    @Override
-    public NewSourceSystemRecord value1(String value) {
-        setId(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord value2(String value) {
-        setName(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord value3(String value) {
-        setEndpoint(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord value4(String value) {
-        setDescription(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord value5(Instant value) {
-        setCreated(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord value6(Instant value) {
-        setDeleted(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord value7(String value) {
-        setMappingId(value);
-        return this;
-    }
-
-    @Override
-    public NewSourceSystemRecord values(String value1, String value2, String value3, String value4, Instant value5, Instant value6, String value7) {
-        value1(value1);
-        value2(value2);
-        value3(value3);
-        value4(value4);
-        value5(value5);
-        value6(value6);
-        value7(value7);
-        return this;
-    }
-
-    // -------------------------------------------------------------------------
-    // Constructors
-    // -------------------------------------------------------------------------
-
-    /**
-     * Create a detached NewSourceSystemRecord
-     */
-    public NewSourceSystemRecord() {
-        super(NewSourceSystem.NEW_SOURCE_SYSTEM);
-    }
-
-    /**
-     * Create a detached, initialised NewSourceSystemRecord
-     */
-    public NewSourceSystemRecord(String id, String name, String endpoint, String description, Instant created, Instant deleted, String mappingId) {
-        super(NewSourceSystem.NEW_SOURCE_SYSTEM);
-
-        setId(id);
-        setName(name);
-        setEndpoint(endpoint);
-        setDescription(description);
-        setCreated(created);
-        setDeleted(deleted);
-        setMappingId(mappingId);
-    }
-}
