+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/DefaultCatalog.java
@@ -6,6 +6,7 @@ package eu.dissco.orchestration.backend.database.jooq;
 
 import java.util.Arrays;
 import java.util.List;
+
 import org.jooq.Schema;
 import org.jooq.impl.CatalogImpl;
