--- src/main/java/eu/dissco/orchestration/backend/database/jooq/Indexes.java
@@ -1,26 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package eu.dissco.orchestration.backend.database.jooq;
-
-
-import eu.dissco.orchestration.backend.database.jooq.tables.Handles;
-import org.jooq.Index;
-import org.jooq.OrderField;
-import org.jooq.impl.DSL;
-import org.jooq.impl.Internal;
-
-
-/**
- * A class modelling indexes of tables in public.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Indexes {
-
-    // -------------------------------------------------------------------------
-    // INDEX definitions
-    // -------------------------------------------------------------------------
-
-    public static final Index DATAINDEX = Internal.createIndex(DSL.name("dataindex"), Handles.HANDLES, new OrderField[] { Handles.HANDLES.DATA }, false);
-    public static final Index HANDLEINDEX = Internal.createIndex(DSL.name("handleindex"), Handles.HANDLES, new OrderField[] { Handles.HANDLES.HANDLE }, false);
-}
