+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/Public.java
@@ -4,11 +4,12 @@
 package eu.dissco.orchestration.backend.database.jooq;
 
 
-import eu.dissco.orchestration.backend.database.jooq.tables.Handles;
 import eu.dissco.orchestration.backend.database.jooq.tables.NewMapping;
 import eu.dissco.orchestration.backend.database.jooq.tables.NewSourceSystem;
+
 import java.util.Arrays;
 import java.util.List;
+
 import org.jooq.Catalog;
 import org.jooq.Table;
 import org.jooq.impl.SchemaImpl;
@@ -26,11 +27,6 @@ public class Public extends SchemaImpl {
      * The reference instance of <code>public</code>
      */
     public static final Public PUBLIC = new Public();
-
-    /**
-     * The table <code>public.handles</code>.
-     */
-    public final Handles HANDLES = Handles.HANDLES;
 
     /**
      * The table <code>public.new_mapping</code>.
@@ -58,7 +54,6 @@ public class Public extends SchemaImpl {
     @Override
     public final List<Table<?>> getTables() {
         return Arrays.<Table<?>>asList(
-            Handles.HANDLES,
             NewMapping.NEW_MAPPING,
             NewSourceSystem.NEW_SOURCE_SYSTEM);
     }
