+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/Tables.java
@@ -4,7 +4,6 @@
 package eu.dissco.orchestration.backend.database.jooq;
 
 
-import eu.dissco.orchestration.backend.database.jooq.tables.Handles;
 import eu.dissco.orchestration.backend.database.jooq.tables.NewMapping;
 import eu.dissco.orchestration.backend.database.jooq.tables.NewSourceSystem;
 
@@ -14,11 +13,6 @@ import eu.dissco.orchestration.backend.database.jooq.tables.NewSourceSystem;
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class Tables {
-
-    /**
-     * The table <code>public.handles</code>.
-     */
-    public static final Handles HANDLES = Handles.HANDLES;
 
     /**
      * The table <code>public.new_mapping</code>.
