+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/tables/NewMapping.java
@@ -7,15 +7,17 @@ package eu.dissco.orchestration.backend.database.jooq.tables;
 import eu.dissco.orchestration.backend.database.jooq.Keys;
 import eu.dissco.orchestration.backend.database.jooq.Public;
 import eu.dissco.orchestration.backend.database.jooq.tables.records.NewMappingRecord;
+
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
+
 import org.jooq.Field;
 import org.jooq.ForeignKey;
 import org.jooq.JSONB;
 import org.jooq.Name;
 import org.jooq.Record;
-import org.jooq.Row8;
+import org.jooq.Row9;
 import org.jooq.Schema;
 import org.jooq.Table;
 import org.jooq.TableField;
@@ -86,6 +88,11 @@ public class NewMapping extends TableImpl<NewMappingRecord> {
      * The column <code>public.new_mapping.deleted</code>.
      */
     public final TableField<NewMappingRecord, Instant> DELETED = createField(DSL.name("deleted"), SQLDataType.INSTANT, this, "");
+
+    /**
+     * The column <code>public.new_mapping.sourcedatastandard</code>.
+     */
+    public final TableField<NewMappingRecord, String> SOURCEDATASTANDARD = createField(DSL.name("sourcedatastandard"), SQLDataType.VARCHAR.nullable(false).defaultValue(DSL.field("'dwca'::character varying", SQLDataType.VARCHAR)), this, "");
 
     private NewMapping(Name alias, Table<NewMappingRecord> aliased) {
         this(alias, aliased, null);
@@ -162,11 +169,11 @@ public class NewMapping extends TableImpl<NewMappingRecord> {
     }
 
     // -------------------------------------------------------------------------
-    // Row8 type methods
+    // Row9 type methods
     // -------------------------------------------------------------------------
 
     @Override
-    public Row8<String, Integer, String, String, JSONB, Instant, String, Instant> fieldsRow() {
-        return (Row8) super.fieldsRow();
+    public Row9<String, Integer, String, String, JSONB, Instant, String, Instant, String> fieldsRow() {
+        return (Row9) super.fieldsRow();
     }
 }
