+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/tables/NewSourceSystem.java
@@ -7,9 +7,11 @@ package eu.dissco.orchestration.backend.database.jooq.tables;
 import eu.dissco.orchestration.backend.database.jooq.Keys;
 import eu.dissco.orchestration.backend.database.jooq.Public;
 import eu.dissco.orchestration.backend.database.jooq.tables.records.NewSourceSystemRecord;
+
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
+
 import org.jooq.Field;
 import org.jooq.ForeignKey;
 import org.jooq.Name;
