+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/tables/records/NewMappingRecord.java
@@ -5,12 +5,14 @@ package eu.dissco.orchestration.backend.database.jooq.tables.records;
 
 
 import eu.dissco.orchestration.backend.database.jooq.tables.NewMapping;
+
 import java.time.Instant;
+
 import org.jooq.Field;
 import org.jooq.JSONB;
 import org.jooq.Record2;
-import org.jooq.Record8;
-import org.jooq.Row8;
+import org.jooq.Record9;
+import org.jooq.Row9;
 import org.jooq.impl.UpdatableRecordImpl;
 
 
@@ -18,7 +20,7 @@ import org.jooq.impl.UpdatableRecordImpl;
  * This class is generated by jOOQ.
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> implements Record8<String, Integer, String, String, JSONB, Instant, String, Instant> {
+public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> implements Record9<String, Integer, String, String, JSONB, Instant, String, Instant, String> {
 
     private static final long serialVersionUID = 1L;
 
@@ -134,6 +136,20 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
         return (Instant) get(7);
     }
 
+    /**
+     * Setter for <code>public.new_mapping.sourcedatastandard</code>.
+     */
+    public void setSourcedatastandard(String value) {
+        set(8, value);
+    }
+
+    /**
+     * Getter for <code>public.new_mapping.sourcedatastandard</code>.
+     */
+    public String getSourcedatastandard() {
+        return (String) get(8);
+    }
+
     // -------------------------------------------------------------------------
     // Primary key information
     // -------------------------------------------------------------------------
@@ -144,17 +160,17 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
     }
 
     // -------------------------------------------------------------------------
-    // Record8 type implementation
+    // Record9 type implementation
     // -------------------------------------------------------------------------
 
     @Override
-    public Row8<String, Integer, String, String, JSONB, Instant, String, Instant> fieldsRow() {
-        return (Row8) super.fieldsRow();
+    public Row9<String, Integer, String, String, JSONB, Instant, String, Instant, String> fieldsRow() {
+        return (Row9) super.fieldsRow();
     }
 
     @Override
-    public Row8<String, Integer, String, String, JSONB, Instant, String, Instant> valuesRow() {
-        return (Row8) super.valuesRow();
+    public Row9<String, Integer, String, String, JSONB, Instant, String, Instant, String> valuesRow() {
+        return (Row9) super.valuesRow();
     }
 
     @Override
@@ -198,6 +214,11 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
     }
 
     @Override
+    public Field<String> field9() {
+        return NewMapping.NEW_MAPPING.SOURCEDATASTANDARD;
+    }
+
+    @Override
     public String component1() {
         return getId();
     }
@@ -238,6 +259,11 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
     }
 
     @Override
+    public String component9() {
+        return getSourcedatastandard();
+    }
+
+    @Override
     public String value1() {
         return getId();
     }
@@ -275,6 +301,11 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
     @Override
     public Instant value8() {
         return getDeleted();
+    }
+
+    @Override
+    public String value9() {
+        return getSourcedatastandard();
     }
 
     @Override
@@ -326,7 +357,13 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
     }
 
     @Override
-    public NewMappingRecord values(String value1, Integer value2, String value3, String value4, JSONB value5, Instant value6, String value7, Instant value8) {
+    public NewMappingRecord value9(String value) {
+        setSourcedatastandard(value);
+        return this;
+    }
+
+    @Override
+    public NewMappingRecord values(String value1, Integer value2, String value3, String value4, JSONB value5, Instant value6, String value7, Instant value8, String value9) {
         value1(value1);
         value2(value2);
         value3(value3);
@@ -335,6 +372,7 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
         value6(value6);
         value7(value7);
         value8(value8);
+        value9(value9);
         return this;
     }
 
@@ -352,7 +390,7 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
     /**
      * Create a detached, initialised NewMappingRecord
      */
-    public NewMappingRecord(String id, Integer version, String name, String description, JSONB mapping, Instant created, String creator, Instant deleted) {
+    public NewMappingRecord(String id, Integer version, String name, String description, JSONB mapping, Instant created, String creator, Instant deleted, String sourcedatastandard) {
         super(NewMapping.NEW_MAPPING);
 
         setId(id);
@@ -363,5 +401,6 @@ public class NewMappingRecord extends UpdatableRecordImpl<NewMappingRecord> impl
         setCreated(created);
         setCreator(creator);
         setDeleted(deleted);
+        setSourcedatastandard(sourcedatastandard);
     }
 }
