+++ src/main/java/eu/dissco/orchestration/backend/database/jooq/tables/records/NewSourceSystemRecord.java
@@ -5,7 +5,9 @@ package eu.dissco.orchestration.backend.database.jooq.tables.records;
 
 
 import eu.dissco.orchestration.backend.database.jooq.tables.NewSourceSystem;
+
 import java.time.Instant;
+
 import org.jooq.Field;
 import org.jooq.Record1;
 import org.jooq.Record7;
